refactor:refactor http util
